+++ source/EngineGpuKernels/Genome.cuh
@@ -2,12 +2,14 @@
 
 #include <cstdint>
 
+#include <cuda_fp16.h>
+
 #include <EngineInterface/CellTypeConstants.h>
 #include <EngineInterface/SimulationParametersTypes.h>
 
 struct NeuralNetworkGenome
 {
-    float weights[MAX_CHANNELS * MAX_CHANNELS];
+    half weights[MAX_CHANNELS * MAX_CHANNELS];  // FP16 for tensor cores
     float biases[MAX_CHANNELS];
     ActivationFunction activationFunctions[MAX_CHANNELS];
     float connectionWeights[MAX_OBJECT_CONNECTIONS];
@@ -239,7 +241,7 @@ struct MemoryGenome
     MemoryModeDataGenome modeData;
 
     uint8_t numSignalEntries;
-    uint8_t channelBitMask;
+    uint16_t channelBitMask;
     SignalEntryGenome* signalEntries;  // Pointer to heap memory
 };
